feat: Dynamic List component
Component that generates a dynamic list of items. It takes in a type, list, and updateList as props. The type defines the type of items, the list holds the items' data, and the updateList function is used to modify the list. The component renders an editable list where each item can be modified or removed. Empty items are automatically managed: when an item is edited, empty items are removed, and a new empty item is added if the last item isn't empty.